requirements.txt
streamlit
pillow
ollama